requirements.txt
numpy
opencv-python
flask
pypng
jsonpickle
dlib
opencv-python
numpy
scikit-learn
Pillow
